pyproject.toml
[tool.poetry]
name = "rai"
version = "0.1.0"
description = "your_description_here"
readme = "README.md"
authors = ["Maciej Majek <maciej.majek@robotec.ai>", "Bartłomiej Boczek <bartlomiej.boczek@robotec.ai>"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
package-mode = false
[tool.poetry.dependencies]
python = "^3.10, <3.13"
requests = "^2.32.2"
pre-commit = "^3.7.0"
openai = "^1.23.3"
coloredlogs = "^15.0.1"
opencv-python = "^4.9.0.80"
markdown = "^3.6"
boto3 = "^1.34.98"
tqdm = "^4.66.4"
pytest = "^8.2.0"
xxhash = "^3.4.1"
optuna = "^3.6.1"
langchain = "^0.2.3"
langchain-aws = "^0.1.7"
langchain-openai = "^0.1.8"
langchain-community = "^0.2.4"
transforms3d = "^0.4.1"
langgraph = "^0.0.66"
tabulate = "^0.9.0"
lark = "^1.1.9"
langfuse = "^2.36.1"
sounddevice = "^0.4.7"
netifaces = "^0.11.0"
redis = "^5.0.7"
openai-whisper = "^20231117"
scipy = "^1.14.0"
torchaudio = "^2.3.1"
elevenlabs = "^1.4.1"
faiss-cpu = "^1.8.0.post1"
rich = "^13.7.1"
docx2txt = "^0.8"
pypdf = "^4.2.0"
langchain-ollama = "^0.1.1"
piper-tts = "^1.2.0"

streamlit = "^1.37.1"
[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"


[tool.poetry.group.gdino]
optional = true

[tool.poetry.group.gdino.dependencies]
pip = "^24.1.2"
torch = "^2.3.1"
torchvision = "^0.18.1"
rf-groundingdino = "^0.2.0"


[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
markers = [
    "billable: marks test as billable  (deselect with '-m \"not billable\"')",
]
addopts = "-m 'not billable' --ignore=src"

[tool.poetry.scripts]
create_rai_ws = "rai.cli.rai_cli:create_rai_ws"
parse_whoami_package = "rai.cli.rai_cli:parse_whoami_package"
